feat: cleaup stuck partitionReassignments
Sometimes, an active partition reassignment goes into a limbo state
due to the destination and source brokers going offline at the same time.
When this happens, there will be a partitionReassignment stuck in kafka
until it is maually cleared- due to this, CC stops reacting to any
anomalies/broker failures/etc.

This commit is for detecting and fixing such stuck active partitionReassignments.